验证通过

+++ IDC_TestSystem/SI-PDU/si_rtu/si_rtu_sent.cpp
@@ -91,7 +91,7 @@ static int rtu_sent_buff(uchar addr, uchar line, uchar *buf)
   * 出口参数：ptr -> 缓冲区
   * 返回值：打包后的长度
   */
-static int rtu_sent_setCmd(uchar addr, uchar reg, ushort value, uchar *buf)
+static int rtu_sent_setCmd(uchar addr, ushort reg, ushort value, uchar *buf)
 {
     static SI_Rtu_Sent msg;
     static QMutex mutex; // 互拆锁
@@ -113,7 +113,7 @@ int SI_RtuSent::sentDataBuff(uchar addr, uchar line, uchar *buf)
     return rtu_sent_buff(addr, line, buf);
 }
 
-int SI_RtuSent::sentCmdBuff(uchar addr, uchar reg, ushort value, uchar *buf)
+int SI_RtuSent::sentCmdBuff(uchar addr, ushort reg, ushort value, uchar *buf)
 {
     return rtu_sent_setCmd(addr, reg, value, buf);
 }
