+++ IDC_TestSystem/SI-PDU/si_rtu/si_rtu_sent.h
@@ -20,7 +20,7 @@ class SI_RtuSent
 {
 public:
     int sentDataBuff(uchar addr, uchar line, uchar *buf);
-    int sentCmdBuff(uchar addr, uchar reg, ushort value, uchar *buf);
+    int sentCmdBuff(uchar addr, ushort reg, ushort value, uchar *buf);
 };
 
 ushort si_rtu_crc(uchar *buf, int len);
